Работа с датами. Добавил функцию remained_time
